refactored the rule out.
